Main now handles authenticated viewing, restructures hierachy to create View, which contains the three visible sections, nav, main, and footer
