+++ src/components/Landing.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+import '../resources/css/navbar.css';
+
+export default class Navbar extends Component {
+    render() {
+        return (
+            <div className="content">
+                <p> Please Login </p>
+            </div>
+        );
+    }
+}
