+++ src/utils/functionalComponent/Main.js
@@ -1,0 +1,22 @@
+import React from "react";
+import Dashboard from "./Dashboard";
+import Landing from "../../components/Landing"
+import { useAuth } from '../useAuth';
+
+function Main() {
+    const firebase = useAuth();
+
+    return (
+        <main className="content">
+            {firebase.user ?
+                <Dashboard />
+                :
+                <Landing />
+            }
+        </main>
+    );
+}
+
+export default Main;
+
+
