+++ src/utils/functionalComponent/View.js
@@ -1,0 +1,18 @@
+import React from "react";
+import Navbar from '../../components/Navbar';
+import Main from './Main';
+import Footer from '../../components/Footer';
+
+function View() {
+    return (
+        <React.Fragment>
+            <Navbar />
+            <Main />
+            <Footer />
+        </React.Fragment>
+    );
+}
+
+export default View;
+
+
